requirements.txt
fastapi>=0.116,<1
uvicorn[standard]>=0.35,<1
python-dotenv>=1.1,<2
pydantic>=2.11,<3
openai>=1.100,<2
httpx>=0.28,<1
pytest>=9,<10
